Trying out rally db ensure instead of recreate
Idea is to be able to reinstall a server from scratch, use
install_rally.sh to setup all the things, and be able to retain the data
in the SQL database.

If I'm running performance benchmarks I'd like to retain them in some
place where I can go back later.

 - #CCCP-2759